requirements.txt
asgiref==3.5.2
Django==4.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6